+++ src/vec3.rs
@@ -4,10 +4,10 @@ use core::ops::{Add, Div, DivAssign, Mul, Neg, Sub};
 use crate::products::{CrossProduct, DotProduct};
 
 #[derive(Copy, Clone, Debug, PartialEq)]
-pub struct Vec3(pub f32, pub f32, pub f32);
+pub struct Vec3(pub f64, pub f64, pub f64);
 
 impl DotProduct for Vec3 {
-	type Output = f32;
+	type Output = f64;
 	fn dot<T: Borrow<Self>>(&self, other: T) -> Self::Output {
 		let other: &Self = other.borrow();
 		(self.0 * other.0) + (self.1 * other.1) + (self.2 * other.2)
@@ -26,11 +26,11 @@ impl CrossProduct for Vec3 {
 }
 
 impl Vec3 {
-	pub fn mag(&self) -> f32 {
-		(self.0.powi(2) as f32 + self.1.powi(2) as f32 + self.2.powi(2) as f32).sqrt()
+	pub fn mag(&self) -> f64 {
+		(self.0.powi(2) as f64 + self.1.powi(2) as f64 + self.2.powi(2) as f64).sqrt()
 	}
 
-	pub fn from_components_with_mag(components: (f32, f32, f32), magnitude: f32) -> Self {
+	pub fn from_components_with_mag(components: (f64, f64, f64), magnitude: f64) -> Self {
 		let unit: Vec3 = Vec3(components.0, components.1, components.2).unit();
 		unit * magnitude
 	}
@@ -56,10 +56,10 @@ impl Add<Vec3> for Vec3 {
 	}
 }
 
-impl Div<f32> for Vec3 {
+impl Div<f64> for Vec3 {
 	type Output = Vec3;
 
-	fn div(self, scalar: f32) -> Self::Output {
+	fn div(self, scalar: f64) -> Self::Output {
 		Vec3(self.0 / scalar, self.1 / scalar, self.2 / scalar)
 	}
 }
@@ -72,7 +72,7 @@ impl Sub<Vec3> for Vec3 {
 	}
 }
 
-impl Mul<Vec3> for f32 {
+impl Mul<Vec3> for f64 {
 	type Output = Vec3;
 
 	fn mul(self, vec: Vec3) -> Self::Output {
@@ -80,16 +80,16 @@ impl Mul<Vec3> for f32 {
 	}
 }
 
-impl Mul<f32> for Vec3 {
+impl Mul<f64> for Vec3 {
 	type Output = Vec3;
 
-	fn mul(self, scalar: f32) -> Self::Output {
+	fn mul(self, scalar: f64) -> Self::Output {
 		Vec3(scalar * self.0, scalar * self.1, scalar * self.2)
 	}
 }
 
-impl DivAssign<f32> for Vec3 {
-	fn div_assign(&mut self, scalar: f32) {
+impl DivAssign<f64> for Vec3 {
+	fn div_assign(&mut self, scalar: f64) {
 		self.0 /= scalar;
 		self.1 /= scalar;
 		self.2 /= scalar;
